+++ R/snippet.R
@@ -76,6 +76,8 @@
 #' (default) for non-standard tunings only, \code{TRUE} or \code{FALSE} for
 #' force on or off completely.
 #' @param midi logical, also output an corresponding MIDI file.
+#' @param colors a named list of LilyPond element color global overrides. See
+#' \code{lilypond} for details.
 #' @param transparent logical, transparent background, png only.
 #'
 #' @return nothing returned; a file is written.
@@ -108,54 +110,55 @@
 #' }
 render_music <- function(music, file, staff = "treble", tuning = "standard",
                          no_tab = TRUE, string_names = NULL, header = NULL,
-                         paper = NULL, midi = FALSE, transparent = FALSE){
+                         paper = NULL, midi = FALSE, colors = NULL,
+                         transparent = FALSE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(tuning, music_staff = staff, no_tab = no_tab) %>%
     score() %>%
     tab(file, ktt[1], ktt[2], ktt[3], header, paper, string_names, TRUE, midi,
-        FALSE, NULL, TRUE, transparent, FALSE)
+        FALSE, colors, TRUE, transparent, FALSE)
 }
 
 #' @export
 #' @rdname render_music
 render_music_tc <- function(music, file, header = NULL, paper = NULL,
-                            midi = FALSE, transparent = FALSE){
+                            midi = FALSE, colors = NULL, transparent = FALSE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(music_staff = "treble", no_tab = TRUE) %>%
     score() %>%
     tab(file, ktt[1], ktt[2], ktt[3], header, paper, FALSE, TRUE, midi, FALSE,
-        NULL, TRUE, transparent, FALSE)
+        colors, TRUE, transparent, FALSE)
 }
 
 #' @export
 #' @rdname render_music
 render_music_bc <- function(music, file, header = NULL, paper = NULL,
-                            midi = FALSE, transparent = FALSE){
+                            midi = FALSE, colors = NULL, transparent = FALSE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(music_staff = "bass", no_tab = TRUE) %>%
     score() %>%
     tab(file, ktt[1], ktt[2], ktt[3], header, paper, FALSE, TRUE, midi, FALSE,
-        NULL, TRUE, transparent, FALSE)
+        colors, TRUE, transparent, FALSE)
 }
 
 #' @export
 #' @rdname render_music
 render_music_tab <- function(music, file, staff = NA, tuning = "standard",
                              string_names = NULL, header = NULL, paper = NULL,
-                             midi = FALSE, transparent = FALSE){
+                             midi = FALSE, colors = NULL, transparent = FALSE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(tuning, music_staff = staff) %>%
     score() %>%
     tab(file, ktt[1], ktt[2], ktt[3], header, paper, string_names, TRUE, midi,
-        FALSE, NULL, TRUE, transparent, FALSE)
+        FALSE, colors, TRUE, transparent, FALSE)
 }
 
 #' @export
@@ -163,28 +166,28 @@ render_music_tab <- function(music, file, staff = NA, tuning = "standard",
 render_music_guitar <- function(music, file, tuning = "standard",
                                 string_names = NULL, header = NULL,
                                 paper = NULL, midi = FALSE,
-                                transparent = FALSE){
+                                colors = NULL, transparent = FALSE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(tuning, music_staff = "treble_8") %>%
     score() %>%
     tab(file, ktt[1], ktt[2], ktt[3], header, paper, string_names, TRUE, midi,
-        FALSE, NULL, TRUE, transparent, FALSE)
+        FALSE, colors, TRUE, transparent, FALSE)
 }
 
 #' @export
 #' @rdname render_music
 render_music_bass <- function(music, file, tuning = "bass",
                               string_names = NULL, header = NULL, paper = NULL,
-                              midi = FALSE, transparent = FALSE){
+                              midi = FALSE, colors = NULL, transparent = FALSE){
   ktt <- .ktt(music)
   paper <- .paper_snippet(paper)
   phrase(music) %>%
     track(tuning, music_staff = "bass_8") %>%
     score() %>%
     tab(file, ktt[1], ktt[2], ktt[3], header, paper, string_names, TRUE, midi,
-        FALSE, NULL, TRUE, transparent, FALSE)
+        FALSE, colors, TRUE, transparent, FALSE)
 }
 
 .paper_snippet <- function(x){